refactor: 컴포넌트 및 함수 분리
